Fix: XSLT document('') function in polyfill
This commit fixes a bug where the XSLT 'document('')' function
did not correctly refer to the stylesheet itself, instead attempting
to load the calling JavaScript file.

The root cause was that the stylesheet's base URI was not being
passed to the underlying libxslt library when parsing the stylesheet
from a string.

Closes: https://github.com/mfreed7/xslt_polyfill/issues/11